+++ spec/DefinitionSpec.php
@@ -13,6 +13,7 @@ namespace spec\Xabbuh\XApi\Model;
 
 use PhpSpec\ObjectBehavior;
 use Xabbuh\XApi\Model\Definition;
+use Xabbuh\XApi\Model\Extensions;
 use Xabbuh\XApi\Model\LanguageMap;
 
 class DefinitionSpec extends ObjectBehavior
@@ -90,6 +91,18 @@ class DefinitionSpec extends ObjectBehavior
         $definition->getMoreInfo()->shouldReturn('https://github.com/adlnet/xAPI_LRS_Test');
     }
 
+    public function it_returns_a_new_instance_with_extensions()
+    {
+        $extensions = new Extensions(array('http://id.tincanapi.com/extension/subject' => 'Conformance Testing'));
+        $definition = $this->withExtensions($extensions);
+
+        $this->getExtensions()->shouldBeNull();
+
+        $definition->shouldNotBe($this);
+        $definition->shouldBeAnInstanceOf(get_class($this->getWrappedObject()));
+        $definition->getExtensions()->shouldReturn($extensions);
+    }
+
     function it_is_different_when_names_are_omitted_and_other_definition_contains_an_empty_list_of_names()
     {
         $this->equals(new Definition(new LanguageMap()))->shouldReturn(false);
@@ -98,6 +111,33 @@ class DefinitionSpec extends ObjectBehavior
     function it_is_different_when_descriptions_are_omitted_and_other_definition_contains_an_empty_list_of_descriptions()
     {
         $this->equals(new Definition(null, new LanguageMap()))->shouldReturn(false);
+    }
+
+    function it_is_not_equal_to_other_definition_if_only_this_definition_has_extensions()
+    {
+        $this->beConstructedWith(null, null, null, null, new Extensions(array('http://id.tincanapi.com/extension/subject' => 'Conformance Testing')));
+
+        $this->equals($this->createEmptyDefinition())->shouldReturn(false);
+    }
+
+    function it_is_not_equal_to_other_definition_if_only_the_other_definition_has_extensions()
+    {
+        $this->beConstructedWith();
+
+        $definition = $this->createEmptyDefinition();
+        $definition = $definition->withExtensions(new Extensions(array('http://id.tincanapi.com/extension/subject' => 'Conformance Testing')));
+
+        $this->equals($definition)->shouldReturn(false);
+    }
+
+    function it_is_not_equal_to_other_definition_if_extensions_are_not_equal()
+    {
+        $this->beConstructedWith(null, null, null, null, new Extensions(array('http://id.tincanapi.com/extension/subject' => 'Conformance Testing')));
+
+        $definition = $this->createEmptyDefinition();
+        $definition = $definition->withExtensions(new Extensions(array('http://id.tincanapi.com/extension/topic' => 'Conformance Testing')));
+
+        $this->equals($definition)->shouldReturn(false);
     }
 
     protected function createEmptyDefinition()