+++ src/Interaction/SequencingInteractionDefinition.php
@@ -12,6 +12,7 @@
 namespace Xabbuh\XApi\Model\Interaction;
 
 use Xabbuh\XApi\Model\Definition;
+use Xabbuh\XApi\Model\Extensions;
 use Xabbuh\XApi\Model\LanguageMap;
 
 /**
@@ -28,12 +29,13 @@ final class SequencingInteractionDefinition extends InteractionDefinition
      * @param LanguageMap|null            $description
      * @param string|null                 $type
      * @param string|null                 $moreInfo
+     * @param Extensions|null             $extensions
      * @param string[]|null               $correctResponsesPattern
      * @param InteractionComponent[]|null $choices
      */
-    public function __construct(LanguageMap $name = null, LanguageMap $description = null, $type = null, $moreInfo = null, array $correctResponsesPattern = null, array $choices = null)
+    public function __construct(LanguageMap $name = null, LanguageMap $description = null, $type = null, $moreInfo = null, Extensions $extensions = null, array $correctResponsesPattern = null, array $choices = null)
     {
-        parent::__construct($name, $description, $type, $moreInfo, $correctResponsesPattern);
+        parent::__construct($name, $description, $type, $moreInfo, $extensions, $correctResponsesPattern);
 
         $this->choices = $choices;
     }